requirements.txt
yfinance
pandas
matplotlib
python-docx
numpy